vision using complex shit
sooo...
Layers:
BehindMask = stuff only shown when in view ie zombies and soldiers.
Mask = a special layer which renders a mask for the BehindMask layer
Blackout = a layer which sits higher and just blacks out the screen a bit.. uses the Inverse of the Mask layer to pick where to render
MapObjects = Anything we want to stop vision (checked against raycasts)

Roughly how it works...
Fire out a load of raycasts and create a mesh for the area we can see (at 1m above ground so will go over smaller objects)

Render everything except zombies and the special layers

Render the Mask layer objects onto as a stencil

Then use that stencil to work out which bits of BehindMask and BlackOut to render.